Fixing Bug https://jira.continuuity.com/browse/ENG-3792 - Empty Log Directories not getting deleted

+++ watchdog/src/main/java/com/continuuity/logging/write/LogCleanup.java
@@ -6,7 +6,6 @@ package com.continuuity.logging.write;
 
 import com.google.common.base.Throwables;
 import com.google.common.collect.Sets;
-import org.apache.hadoop.ipc.RemoteException;
 import org.apache.twill.filesystem.Location;
 import org.apache.twill.filesystem.LocationFactory;
 import org.slf4j.Logger;
@@ -22,13 +21,11 @@ public final class LogCleanup implements Runnable {
   private static final Logger LOG = LoggerFactory.getLogger(LogCleanup.class);
 
   private final FileMetaDataManager fileMetaDataManager;
-  private final LocationFactory locationFactory;
   private final Location logBaseDir;
   private final long retentionDurationMs;
 
   public LogCleanup(LocationFactory locationFactory, FileMetaDataManager fileMetaDataManager, Location logBaseDir,
              long retentionDurationMs) {
-    this.locationFactory = locationFactory;
     this.fileMetaDataManager = fileMetaDataManager;
     this.logBaseDir = LocationUtils.normalize(locationFactory, logBaseDir);
     this.retentionDurationMs = retentionDurationMs;
@@ -52,8 +49,7 @@ public final class LogCleanup implements Runnable {
                                                 LOG.info(String.format("Deleting log file %s", location.toURI()));
                                                 location.delete();
                                               }
-
-                                              parentDirs.add(LocationUtils.getParent(locationFactory, location));
+                                              parentDirs.add(LocationUtils.getParent(location));
                                             } catch (IOException e) {
                                               LOG.error(
                                                 String.format("Got exception when deleting path %s",
@@ -79,7 +75,6 @@ public final class LogCleanup implements Runnable {
    * @param dir dir to be deleted.
    */
   void deleteEmptyDir(Location dir) {
-    dir = LocationUtils.normalize(locationFactory, dir);
     LOG.debug("Got path {}", dir.toURI());
 
     // Don't delete a dir if it is equal to or a parent of logBaseDir
@@ -90,19 +85,16 @@ public final class LogCleanup implements Runnable {
     }
 
     try {
-      if (dir.delete()) {
+      if (dir.list().isEmpty() && dir.delete()) {
         LOG.info("Deleted empty dir {}", dir.toURI());
 
         // See if parent dir is empty, and needs deleting
-        Location parent = LocationUtils.getParent(locationFactory, dir);
+        Location parent = LocationUtils.getParent(dir);
         LOG.debug("Deleting parent dir {}", parent);
         deleteEmptyDir(parent);
       } else {
         LOG.debug("Not deleting non-dir or non-empty dir {}", dir.toURI());
       }
-    } catch (RemoteException e) {
-      // Don't log non-empty dir deletion exception as error, as it is expected.
-      LOG.debug("Got exception while deleting dir {}", dir.toURI(), e);
     } catch (IOException e) {
       LOG.error("Got exception while deleting dir {}", dir.toURI(), e);
     }
